Refactor - relocate authentication.py and typing.py to internal dir.
